SchoolManagement/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-bootstrap5==24.2
django-mssql==1.8
djangorestframework==3.15.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mssql-django==1.5
mysqlclient==2.2.4
pyodbc==5.1.0
pytz==2024.1
PyYAML==6.0.1
render==1.0.0
sqlparse==0.5.0
tzdata==2024.1
